rfqmsg: move encode/decode functions to records.go
