+++ rfqmsg/buy_request.go
@@ -2,9 +2,7 @@ package rfqmsg
 
 import (
 	"crypto/rand"
-	"crypto/sha256"
 	"fmt"
-	"io"
 
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/lightninglabs/taproot-assets/asset"
@@ -12,65 +10,6 @@ import (
 	"github.com/lightningnetwork/lnd/routing/route"
 	"github.com/lightningnetwork/lnd/tlv"
 )
-
-func IdEncoder(w io.Writer, val any, buf *[8]byte) error {
-	if t, ok := val.(*ID); ok {
-		id := [32]byte(*t)
-		return tlv.EBytes32(w, &id, buf)
-	}
-
-	return tlv.NewTypeForEncodingErr(val, "MessageID")
-}
-
-func IdDecoder(r io.Reader, val any, buf *[8]byte, l uint64) error {
-	const idBytesLen = 32
-
-	if typ, ok := val.(*ID); ok {
-		var idBytes [idBytesLen]byte
-
-		err := tlv.DBytes32(r, &idBytes, buf, idBytesLen)
-		if err != nil {
-			return err
-		}
-
-		id := ID(idBytes)
-
-		*typ = id
-		return nil
-	}
-
-	return tlv.NewTypeForDecodingErr(val, "MessageID", l, idBytesLen)
-}
-
-func AssetIdEncoder(w io.Writer, val any, buf *[8]byte) error {
-	if t, ok := val.(**asset.ID); ok {
-		id := [sha256.Size]byte(**t)
-		return tlv.EBytes32(w, &id, buf)
-	}
-
-	return tlv.NewTypeForEncodingErr(val, "assetId")
-}
-
-func AssetIdDecoder(r io.Reader, val any, buf *[8]byte, l uint64) error {
-	const assetIDBytesLen = sha256.Size
-
-	if typ, ok := val.(**asset.ID); ok {
-		var idBytes [assetIDBytesLen]byte
-
-		err := tlv.DBytes32(r, &idBytes, buf, assetIDBytesLen)
-		if err != nil {
-			return err
-		}
-
-		id := asset.ID(idBytes)
-		assetId := &id
-
-		*typ = assetId
-		return nil
-	}
-
-	return tlv.NewTypeForDecodingErr(val, "assetId", l, sha256.Size)
-}
 
 const (
 	// latestBuyRequestVersion is the latest supported buy request wire
